fix(remediation): close extraction and tls deployment follow-through (#4735)
